v1.9 Enhancements

+++ customize.sh
@@ -1,6 +1,6 @@
 # boot mode
 if [ "$BOOTMODE" != true ]; then
-  abort "- Please flash via Magisk Manager only!"
+  abort "- Please flash via Magisk app only!"
 fi
 
 # space
@@ -24,12 +24,23 @@ fi
 SYSTEM=`realpath $MIRROR/system`
 PRODUCT=`realpath $MIRROR/product`
 VENDOR=`realpath $MIRROR/vendor`
-SYSTEM_EXT=`realpath $MIRROR/system/system_ext`
-ODM=`realpath /odm`
-MY_PRODUCT=`realpath /my_product`
+SYSTEM_EXT=`realpath $MIRROR/system_ext`
+if [ -d $MIRROR/odm ]; then
+  ODM=`realpath $MIRROR/odm`
+else
+  ODM=`realpath /odm`
+fi
+if [ -d $MIRROR/my_product ]; then
+  MY_PRODUCT=`realpath $MIRROR/my_product`
+else
+  MY_PRODUCT=`realpath /my_product`
+fi
 
 # optionals
 OPTIONALS=/sdcard/optionals.prop
+if [ ! -f $OPTIONALS ]; then
+  touch $OPTIONALS
+fi
 
 # info
 MODVER=`grep_prop version $MODPATH/module.prop`
@@ -60,13 +71,12 @@ if [ "$BOOTMODE" != true ]; then
   mount -o rw -t auto /dev/block/bootdevice/by-name/metadata /metadata
 fi
 
-# sepolicy.rule
-FILE=$MODPATH/sepolicy.sh
-DES=$MODPATH/sepolicy.rule
-if [ -f $FILE ] && [ "`grep_prop sepolicy.sh $OPTIONALS`" != 1 ]; then
+# sepolicy
+FILE=$MODPATH/sepolicy.rule
+DES=$MODPATH/sepolicy.pfsd
+if [ "`grep_prop sepolicy.sh $OPTIONALS`" == 1 ]\
+&& [ -f $FILE ]; then
   mv -f $FILE $DES
-  sed -i 's/magiskpolicy --live "//g' $DES
-  sed -i 's/"//g' $DES
 fi
 
 # cleaning
@@ -119,18 +129,29 @@ elif [ "`grep_prop permissive.mode $OPTIONALS`" == 2 ]; then
   ui_print " "
 fi
 
+# /priv-app
+if [ ! -d $SYSTEM/priv-app ]; then
+  ui_print "- /system/priv-app is not supported"
+  ui_print "  Moving to /system/app..."
+  mv -f $MODPATH/system/priv-app $MODPATH/system/app
+  ui_print " "
+fi
+
 # function
 extract_lib() {
-  for APPS in $APP; do
+for APPS in $APP; do
+  FILE=`find $MODPATH/system -type f -name $APPS.apk`
+  if [ -f `dirname $FILE`/extract ]; then
+    rm -f `dirname $FILE`/extract
     ui_print "- Extracting..."
-    FILE=`find $MODPATH/system -type f -name $APPS.apk`
-    DIR=`find $MODPATH/system -type d -name $APPS`/lib/$ARCH
+    DIR=`dirname $FILE`/lib/$ARCH
     mkdir -p $DIR
     rm -rf $TMPDIR/*
     unzip -d $TMPDIR -o $FILE $DES
     cp -f $TMPDIR/$DES $DIR
     ui_print " "
-  done
+  fi
+done
 }
 
 # extract
@@ -142,9 +163,9 @@ chmod 0755 $DIR/*
 # version
 APP=HEXEditor
 PKG=com.myprog.hexedit
-CURRENT=`pm list packages --show-versioncode | grep $PKG | sed -n -e "s/package:$PKG versionCode://p"`
+CURRENT=`pm list packages --show-versioncode | grep $PKG | sed "s/package:$PKG versionCode://"`
 NEW=120
-DIR=/system/priv-app/$APP
+DIR=`find /data/adb/modules/"$MODID" -type d -name "$APP"`
 ui_print "- Current app versionCode: $CURRENT"
 ui_print "  New app versionCode: $NEW"
 if [ "$CURRENT" == "$NEW" ]; then
@@ -161,15 +182,6 @@ if [ "$CURRENT" == "$NEW" ]; then
 fi
 ui_print " "
 
-# grant
-FILE=$MODPATH/system/priv-app/$APP/$APP.apk
-if ! pm list package | grep -Eq $PKG; then
-  ui_print "- Granting all runtime permissions..."
-  RES=`pm install -g -i com.android.vending $FILE`
-  RES=`pm uninstall -k $PKG`
-  ui_print " "
-fi
-
 # power save
 FILE=$MODPATH/system/etc/sysconfig/*
 if [ "`grep_prop power.save $OPTIONALS`" == 1 ]; then
@@ -182,14 +194,17 @@ if [ "`grep_prop power.save $OPTIONALS`" == 1 ]; then
   ui_print " "
 fi
 
-# /priv-app
-if [ ! -d $SYSTEM/priv-app ]; then
-  ui_print "- /system/priv-app is not supported"
-  ui_print "  Moving to /system/app..."
-  rm -rf $MODPATH/system/app
-  mv -f $MODPATH/system/priv-app $MODPATH/system/app
-  ui_print " "
-fi
+# function
+hide_oat() {
+for APPS in $APP; do
+  mkdir -p `find $MODPATH/system -type d -name $APPS`/oat
+  touch `find $MODPATH/system -type d -name $APPS`/oat/.replace
+done
+}
+
+# hide
+APP="`ls $MODPATH/system/priv-app` `ls $MODPATH/system/app`"
+hide_oat
